feat: display learning parameters
