+++ web_ui/src/pages/project-details/components/project-models/train-model-dialog/advanced-settings/training/learning-parameters/learning-parameters.component.tsx
@@ -3,7 +3,6 @@
 
 import { useRef } from 'react';
 
-import { Flex } from '@geti/ui';
 import { isEqual } from 'lodash-es';
 
 import { TrainingConfiguration } from '../../../../../../../../core/configurable-parameters/services/configuration.interface';
